Split up sled.rs into multiple files
